requirements.txt
streamlit>=1.28.0
fireworks-ai>=0.7.0
aiohttp>=3.8.4
asyncio>=3.4.3
python-dotenv>=1.0.0